+++ nxdrive/engine/engine.py
@@ -1,7 +1,6 @@
 # coding: utf-8
 import datetime
 import os
-from cookielib import CookieJar
 from logging import getLogger
 from threading import Thread, current_thread
 from time import sleep
@@ -16,7 +15,7 @@ from .queue_manager import QueueManager
 from .watcher.local_watcher import LocalWatcher
 from .watcher.remote_watcher import RemoteWatcher
 from .workers import PairInterrupt, ThreadInterrupt, Worker
-from ..client import LocalClient, FilteredRemote, Remote
+from ..client import FilteredRemote, LocalClient, Remote
 from ..client.common import NotFound, safe_filename
 from ..options import Options
 from ..osi import AbstractOSIntegration
@@ -93,8 +92,6 @@ class Engine(QObject):
         self.timeout = 30
         self._handshake_timeout = 60
         self.manager = manager
-        # Remove remote client cache on proxy update
-        self.manager.proxyUpdated.connect(self.invalidate_client_cache)
 
         self.local_folder = definition.local_folder
         self.local = self.local_cls(self.local_folder)
@@ -207,10 +204,12 @@ class Engine(QObject):
             sleep(1)
         log.debug('Local Folder lock setup completed on %r', path)
 
-    def set_ui(self, value):
-        self._dao.update_config('ui', value)
-        self._ui = value
-        log.debug('UI preferences set to {}'.format(value))
+    def set_ui(self, value, overwrite=True):
+        name = ['ui', 'force_ui']
+        self._dao.update_config(name[overwrite], value)
+        setattr(self, '_' + name[overwrite], value)
+
+        log.debug('{} preferences set to {}'.format(name[overwrite], value))
 
     def release_folder_lock(self):
         log.debug('Local Folder unlocking')
@@ -278,7 +277,7 @@ class Engine(QObject):
             'edit': ('view_documents', 'view_drive_metadata')[edit],
             'token': self.get_remote_token(),
         }
-        return urls.get(self._ui, 'web').format(**infos)
+        return urls.get(self._force_ui or self._ui).format(**infos)
 
     def get_remote_url(self):
         """
@@ -299,7 +298,7 @@ class Engine(QObject):
             'repo': Options.remote_repo,
             'token': self.get_remote_token(),
         }
-        return urls.get(self._ui, 'web').format(**infos)
+        return urls.get(self._force_ui or self._ui).format(**infos)
 
     def is_syncing(self):
         return self._sync_started
@@ -400,7 +399,8 @@ class Engine(QObject):
         self._remote_user = self._dao.get_config('remote_user')
         self._remote_password = self._dao.get_config('remote_password')
         self._remote_token = self._dao.get_config('remote_token')
-        self._ui = self._dao.get_config('ui', default='web')
+        self._ui = self._dao.get_config('ui', default='jsf')
+        self._force_ui = self._dao.get_config('force_ui')
         if self._remote_password is None and self._remote_token is None:
             self.set_invalid_credentials(
                 reason='found no password nor token in engine configuration')
@@ -681,7 +681,6 @@ class Engine(QObject):
             raise ValueError
         self._dao.update_config('remote_token', self._remote_token)
         self.set_invalid_credentials(value=False)
-        self.invalidate_client_cache()
         # In case of a binding
         self._check_root()
         self.start()
@@ -691,7 +690,6 @@ class Engine(QObject):
         self._remote_token = token
         self._dao.update_config('remote_token', self._remote_token)
         self.set_invalid_credentials(value=False)
-        self.invalidate_client_cache()
         self.start()
 
     def init_remote(self):
@@ -703,12 +701,12 @@ class Engine(QObject):
             self.version,
         )
         rkwargs = {
-            'proxies': self.manager.get_proxies(self._server_url),
             'password': self._remote_password,
             'timeout': self.timeout,
             'token': self._remote_token,
             'check_suspended': self.suspend_client,
             'dao': self._dao,
+            'proxy': self.manager.proxy,
         }
         self.remote = self.filtered_remote_cls(*rargs, **rkwargs)
 
@@ -807,12 +805,6 @@ class Engine(QObject):
                 pair = thread.worker.get_current_pair()
                 if pair is not None and pair.id == pair_id:
                     thread.worker.quit()
-
-    @pyqtSlot()
-    def invalidate_client_cache(self):
-        log.debug('Invalidate client cache')
-        # TODO: called on new proxy settings and update_password()
-        self.invalidClientsCache.emit()
 
     def _set_root_icon(self):
         if not self.local.exists('/') or self.local.has_folder_icon('/'):