+++ nxdrive/options.py
@@ -153,7 +153,6 @@ class MetaOptions(type):
         'nofscheck': (False, 'default'),
         'protocol_url': (None, 'default'),
         'proxy_server': (None, 'default'),
-        'proxy_type': (None, 'default'),
         'remote_repo': ('default', 'default'),
         'res_dir': (
             os.path.join(getattr(sys, '_MEIPASS', os.path.dirname(__file__)),
@@ -394,8 +393,9 @@ def server_updater(*args):
                     conf = resp.json()
                 except Exception as exc:
                     log.error('Polling error: %r', exc)
+                    engine.set_ui('jsf', overwrite=False)
                 else:
-                    engine.set_ui(conf.pop('ui'))
+                    engine.set_ui(conf.pop('ui'), overwrite=False)
                     Options.update(conf, setter='server', fail_on_error=True)
                     break