+++ docs/changes.md
@@ -1,14 +1,23 @@
 # dev
 Release date: `2018-??-??`
 
+Changes in command line arguments:
+- Removed `proxy-exceptions`
+- Removed `proxy-type`: Pass the scheme directly in the `proxy-server` url
+
 ### Core
 - [NXDRIVE-825](https://jira.nuxeo.com/browse/NXDRIVE-825): Rely on the Python client for all Nuxeo API calls (**breaking change**)
+- [NXDRIVE-1068](https://jira.nuxeo.com/browse/NXDRIVE-1068): Move proxy support to a dedicated module
 - [NXDRIVE-1201](https://jira.nuxeo.com/browse/NXDRIVE-1201): Adapt Drive for new Trash API behavior
+
+### GUI
+- [NXDRIVE-1183](https://jira.nuxeo.com/browse/NXDRIVE-1183): Make server UI selection smarter
 
 #### Minor changes
 - Jenkins:Added the `PYTEST_ADDOPTS` parameter to the Drive-tests job
 - Jenkins: Removed the `ENABLE_PROFILER` parameter from the Drive-tests job
 - Packaging: Added `nuxeo` 2.0.0
+- Packaging: Updated `pytest` from 3.5.1 to 3.6.0
 
 
 # 3.1.0